+++ Pipelines/Cmdlets/NewAdoStage.cs
@@ -4,10 +4,10 @@ using System.Management.Automation;
 
 namespace ModPosh.Pipelines.Cmdlets
 {
-    [Cmdlet(VerbsCommon.New, "Stage", HelpUri = "https://github.com/mod-posh/Pipelines/blob/v1.6.0.0/docs/New-Stage.md#new-stage")]
+    [Cmdlet(VerbsCommon.New, "AdoStage", HelpUri = "https://github.com/mod-posh/Pipelines/blob/v2.0.0.0/docs/New-AdoStage.md#new-adostage")]
     [OutputType(typeof(Stage))]
     [CmdletBinding(PositionalBinding = true)]
-    public class NewStage : PSCmdlet
+    public class NewAdoStage : PSCmdlet
     {
         [Parameter(Mandatory = true, Position = 0, ParameterSetName = "Default")]
         public string Name { get; set; } = string.Empty;