+++ Pipelines/Tests/CmdletTests/NewAdoPool.Tests.ps1
@@ -3,27 +3,27 @@ BeforeAll {
     # Import-Module 'Path\To\Your\Module.psm1'
 }
 
-Describe 'New-Pool Cmdlet Tests' {
+Describe 'New-AdoPool Cmdlet Tests' {
 
     Context 'Parameter Tests' {
         It 'Should require a Name parameter' {
-            { New-Pool -Name $null } | Should -Throw -ExpectedMessage "Cannot bind argument to parameter 'Name' because it is null."
+            { New-AdoPool -Name $null } | Should -Throw -ExpectedMessage "Cannot bind argument to parameter 'Name' because it is null."
         }
 
         It 'Should accept a Name parameter' {
-            { New-Pool -Name 'ExamplePool' } | Should -Not -Throw
+            { New-AdoPool -Name 'ExamplePool' } | Should -Not -Throw
         }
     }
 
     Context 'Functionality Tests' {
         It 'Should return an object of type Pool' {
-            $result = New-Pool -Name 'ExamplePool'
+            $result = New-AdoPool -Name 'ExamplePool'
             $result | Should -BeOfType ModPosh.Pipelines.Ado.Pool
         }
 
         It 'Can have Demands parameter' {
             $demands = @('demand1', 'demand2')
-            $result = New-Pool -Name 'ExamplePool' -Demands $demands
+            $result = New-AdoPool -Name 'ExamplePool' -Demands $demands
 
             # Check if the number of demands in the result is equal to the input
             $result.Demands.Count | Should -Be $demands.Length