+++ Pipelines/Tests/CmdletTests/NewAdoStage.Tests.ps1
@@ -4,15 +4,15 @@ BeforeAll {
     # or define Mocks if necessary
 }
 
-Describe 'New-Stage Cmdlet Tests' {
+Describe 'New-AdoStage Cmdlet Tests' {
 
     Context 'Parameter Tests' {
         It 'Should require a Name parameter' {
-            { New-Stage -Name $null } | Should -Throw -ExpectedMessage "Cannot bind argument to parameter 'Name' because it is null."
+            { New-AdoStage -Name $null } | Should -Throw -ExpectedMessage "Cannot bind argument to parameter 'Name' because it is null."
         }
 
         It 'Should accept a Name parameter' {
-            { New-Stage -Name 'ExampleStage' } | Should -Not -Throw
+            { New-AdoStage -Name 'ExampleStage' } | Should -Not -Throw
         }
     }
 
@@ -22,31 +22,31 @@ Describe 'New-Stage Cmdlet Tests' {
         }
 
         It 'Should return an object of type Stage' {
-            $result = New-Stage -Name 'ExampleStage'
+            $result = New-AdoStage -Name 'ExampleStage'
             $result | Should -BeOfType ModPosh.Pipelines.Ado.Stage
         }
 
         It 'Can have a DisplayName parameter' {
             $displayName = 'StageDisplayName'
-            $result = New-Stage -Name 'ExampleStage' -DisplayName $displayName
+            $result = New-AdoStage -Name 'ExampleStage' -DisplayName $displayName
             $result.DisplayName | Should -Be $displayName
         }
 
         It 'Can have DependsOn parameter' {
             $dependsOn = @('Stage1', 'Stage2')
-            $result = New-Stage -Name 'ExampleStage' -DependsOn $dependsOn
+            $result = New-AdoStage -Name 'ExampleStage' -DependsOn $dependsOn
             $result.DependsOn | Should -Be $dependsOn
         }
 
         It 'Can have a Condition parameter' {
             $condition = 'success()'
-            $result = New-Stage -Name 'ExampleStage' -Condition $condition
+            $result = New-AdoStage -Name 'ExampleStage' -Condition $condition
             $result.Condition | Should -Be $condition
         }
 
         It 'Can have Variables parameter' {
             $variables = @{ Variable1 = 'Value1'; Variable2 = 'Value2' }
-            $result = New-Stage -Name 'ExampleStage' -Variables $variables
+            $result = New-AdoStage -Name 'ExampleStage' -Variables $variables
 
             # Ensure $result.Variables contains all keys from $variables
             foreach ($key in $variables.Keys) {
@@ -61,7 +61,7 @@ Describe 'New-Stage Cmdlet Tests' {
 }
         It 'Can have Jobs parameter' {
             $jobs = @([ModPosh.Pipelines.Ado.Job]::new(), [ModPosh.Pipelines.Ado.Job]::new())
-            $result = New-Stage -Name 'ExampleStage' -Jobs $jobs
+            $result = New-AdoStage -Name 'ExampleStage' -Jobs $jobs
             $result.Jobs.Count | Should -Be $jobs.Length
         }